requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1.6
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.11
ipython==8.9.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
